Update host_comms.cpp
Fixed function return type for software time

+++ libraries/buggycommon/host_comms.cpp
@@ -223,7 +223,7 @@ AlarmStatus alarm_status() {
     return ALARM_STATUS;
 }
 
-int software_time() {
+float software_time() {
     return SOFT_TIME;
 }
